requirements.txt
asyncio==3.4.3
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
docutils==0.20
idna==3.4
Kivy==2.1.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kivy-garden.graph==0.4.0
kivymd==1.1.1
MarkupSafe==2.1.2
MouseInfo==0.1.3
mysql-connector-python==8.0.33
Pillow==9.5.0
ping3==4.0.4
protobuf==3.20.3
psutil==5.9.5
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.15.1
PyMsgBox==1.0.9
PyMySQL==1.0.3
pynput==1.7.6
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
python-dotenv==1.0.0
pytweening==1.0.7
pywin32==306
requests==2.30.0
six==1.16.0
urllib3==2.0.2
Werkzeug==2.3.4
